Removed unwanted files and added functionalities
