+++ FIRMWARE_ALFHA1_0/actuator_constants.py
@@ -35,9 +35,9 @@ MINIMUM_DUTY_CYCLE = 1000
 SIT_STAND_TRANSFORMATION_PIN_ARRAY = [3, 11, 6, 14]
 SIT_STAND_TRANSFORMATION_PWM_ARRAY = [5000, 2000, 5000, 8000]
 STAND_SIT_TRANSFORMATION_PWM_ARRAY = [2000, 5000, 8000, 5000]
-STAND_FORWARD_TRANSFORMATION_PIN_ARRAY = [2, 10, 5, 13]
-STAND_FORWARD_TRANSFORMATION_PWM_ARRAY = [5000, 8000, 5000, 2000]
-FORWARD_STAND_TRANSFORMATION_PWM_ARRAY = [8000, 5000, 2000, 5000]
+ALL_ARMS_TRANSFORMATION_ARRAY = [30, 90, 150]
+FORWARD_TRANSFORMATION_ARRAY = [60, 70, 80, 110, 120, 130]
+STAND_DEFAULT_HARD = []
 FRONTAL_ALEG = [2, 3, 4]
 FRONTAL_BLEG = [5, 6, 7]
 HIND_CLEG = [10, 11, 12]
