--- firmware/hardware_Calibration/SERVO_SWEEP
@@ -1,18 +1,0 @@
-from time import sleep
-from machine import Pin
-from machine import PWM
-def Control_sweep(pin):
-    pwm = PWM(Pin(pin))    
-    pwm.freq(50)
-    def setServoCycle (position):
-        pwm.duty_u16(position)
-        #pwm1.duty_u16(position)
-        sleep(0.01)
-    for pos in range(5000,9000,50):
-        setServoCycle(pos)
-    for pos in range(9000,5000,-50):
-        setServoCycle(pos)
-
-while True:    
-    a=int(input("Enter the pin: "))
-    Control_sweep(a)
